refactored pool transactions component, output pool transactions on asset details
